api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
python-dotenv==1.0.0
nanoid==2.0.0
langdetect==1.0.9
anthropic==0.25.0
markdown2==2.5.1